New inference: error if eager solving to nothing
Summary:
Type inference sometimes has to resolve a type variable without using subtyping e.g. for method invocation off an object, or "array" update and append. In these situations we "force" solve the type variable, if possible without losing completeness (because the type variable appears only covariantly, contravariantly, or not at all in relevant types). But if the type variable is invariant (appears both covariantly and contravariantly) then we arbitrarily push it to its lower bounds. If there are no lower bounds, then we solve it to `nothing`. It's rare that this actually makes sense: for example, if the expression is a lambda parameter, then assigning it type `nothing` means that the lambda can never be applied. Better to emit an error at the point that the type variable was inspected.

We might in future try and eliminate some uses of `expand_type_and_solve`. This diff puts us in a better place to decide which uses are important, as the error message that we produce gives some indication of what type is expected (e.g. `an object` or `an array or collection`).

Reviewed By: manzyuk

Differential Revision: D14065305

fbshipit-source-id: 7604c147e7a44ed7367ec2d115a75e838a028137