Fix property type-hint ref binding warnings for static properties
Summary:
The interpreter was using isCheckable() to determine whether it should
raise a ref binding warning/error on a static property with a type-hint. This
isn't correct because a type-alias to mixed will return true for
isCheckable(). Instead we should use isMixedResolved() which resolves
type-aliases. The warning/error should only be raised if !isMixedResolved(),
which means its definitely not mixed.

Reviewed By: markw65

Differential Revision: D14088690

fbshipit-source-id: 3983589d740aa634fab158e2366237b72ef28b74