Allow reified type arguments that are not wildcards in is/as checks
Summary:
For is/as tests, the runtime can enforce deeply nested reified type arguments. It also supports wildcards for reified type arguments.

All of the erased generics must be wildcards still, but if we encounter a reified type argument that isn't a wildcard, we run the type visitor on it.

Reviewed By: oulgen

Differential Revision: D14079639

fbshipit-source-id: a9c3770c3c42f74790d0ea29120339f56b8e080e