Back out "[hack] Handle EINTR in process.ml"
Summary:
I believe D14099177 fixes the root cause of the problem that D14036824 was intended to solve. I think it's likely that D14036824 now has no effect, but I'm not confident that this is the case--it's possible that in some rarely-reached edge case, we *did* want this `select` to be interrupted by EINTR, and thus D14036824 has a *negative* effect.

I'm concerned about this because while running `unattended_release.sh` for the release cut last night (logs: P60951580), hh_client stalled indefinitely when we attempted to run `phps XControllerForcePurity --lint-all-controllers` from `www_test.sh`. I discovered this when checking on the progress of the release 90 minutes later. No server was running, and the monitor log (P60953925) indicated that the monitor was waiting on Informant updates:

```
[2019-02-15 01:41:24.093] Not starting first server. Starting will be triggered by informant later.
```

I tried running hh_client myself, but it also just stalled. I then tried `hh --force-dormant-start true` (as suggested [here](https://our.intern.facebook.com/intern/qa/5770/hh-waiting-for-a-server-to-be-started-for-a-long-t)), which did get the monitor to start the server (monitor log: P60951574):

```
[2019-02-15 03:28:14.079] Reinitializing Informant
[2019-02-15 03:28:14.223] About to spawn typechecker daemon. Logs will go to /tmp/hh_server/zSdatazSuserszSjakebaileyzSwww-hg.log
```

I'm concerned that with D14036824, we have started ignoring a meaningful signal, and this stalling is a possible (but infrequent--I haven't been able to reproduce it) consequence.

It's also possible that this stalling is just an unrelated bug which I encountered only by chance.

Reviewed By: ljw1004

Differential Revision: D14100289

fbshipit-source-id: 30bc9b2a68e284a27f9c9f3747623f6eaaa2d356