Remove --list-modes --show (along with ServerCommandTypes.Stream)
Summary: Build (removed in D13741832) was the only one actually streaming anything, with the rest of "streamed" commands easily able to be RPC calls. D13773665 change sone of them, this diff removes the other two that had no recorded usages in Scuba data, and the whole concept of streaming can be removed.

Reviewed By: arxanas

Differential Revision: D13773715

fbshipit-source-id: 2375a76e1e67facafa296420f2741f88b0ec1251