Free code.prof after retranslate-all completes, take 3
Summary:
This is the new attempt at D13946266, with one small change.  When the JIT bails
out and tries to emit an Interp translation, it's not enough to mark the
function as being profiled as done in the previous fix.  That's because there
won't be any profile translation recorded for the corresponding SrcKey, and
therefore the Interp request emitted in code.prof won't be invalidated (even for
functions marked as being profile, we only invalidate the SrcKeys associated
with profile translations).  So, instead, this diff now simply doesn't emit and
Interp request if the translation that was being attempted was a profile one.

This diff also adds a call to mprotect to prevent any access to the profile code
area after it's madvised.

Reviewed By: binliu19

Differential Revision: D14107312

fbshipit-source-id: 27b88768195684062506c8a6dc4034c4fa658e37