Add <<__MockClass>> to the list of known attributes
Summary:
Add the attribute `<<__MockClass>>` to the list of known attributes.

`<<__MockClass>>` is already understood by HHVM. It enables users to inherit from `final` classes for mocking in tests.

This is not sufficient to type check a class using `<<__MockClass>>`. We still complain about inheriting from a final class. Type checking `<<__MockClass>>` would be very difficult to do correctly (e.g. new static, the `this` type), so require the user to write `HH_IGNORE_ERROR [4035]`.

Reviewed By: vsiles

Differential Revision: D13558385

fbshipit-source-id: 4dd24349f1e40134d0a6fb31eb7b59c5ab8baa5c