Support calling values of generic types constrained to function types
Summary:
Surprisingly enough, we currently don't support calling values of generic types bounded by function types from above.  For example, the following code is rejected by the type-checker:
```
<?hh // strict

function test<T as (function(): void)>(T $f): void {
  $f();
}
```
This is because when we type-check function calls, we expect the type of the function to be concrete (possibly a union) and we error out when it is generic.  A common fix, which is used in a few other similar places, is to iterate over the concrete supertypes of the type of the function.  This fix is not perfect.  In particular, we are still going to reject code where there are multiple upper bounds that are function types (see test cases below).

The morally correct way to type-check a function call like, say, `$f($x)`, where `$f : fty` and `$x : ty`, is to introduce fresh type variables `v` and `w` and impose subtyping constraints `fty <: (function(v): w)` and `ty <: v`.  This is what we are aiming at with the new constraint-based type inference.  At the moment, implementation of this approach is complicated by the fact that function types have additional attributes like coroutineness, reactivity etc., which are taken into account by subtyping checks.  We would need to introduce coroutineness, reactivity etc. *variables* so that these attributes could be left undetermined (as variables) in the function type `(function(v): w)` and resolved by solving the constraints implied by subtyping checks.  This is left for future work.

Reviewed By: DavidSnider, andrewjkennedy

Differential Revision: D14131127

fbshipit-source-id: 15abe3eee1def5d517067b5f6ceba74ff06a0ed6