address m_code race in pageletserver
Summary:
We have had an invariant for a long time in the consumer of the pageletserver APIs: if we see a non-zero status code, this batch of pageletserver results will be the last one we receive, because the non-zero status code comes along with closing the pageletserver transport. If we get flushes after a non-zero exit code, we consider that to be broken.

Right now, there is a slight race that can break this invariant, especially when used with `fb_asyncify_pagelet`. It's possible to upgrade a pagelet task to a WaitHandle (and create a new event for it) even though the pagelet transport was already closed. This doesn't cause HHVM to crash, but it does cause us to do an extra empty flush with no contents and a duplicated status code.

We don't have a lock around the `m_code` update, so if `fb_asyncify_pagelet` is in the critical section, we may grab the value of `m_code` but make decisions around new `PageletServerTaskEvent` creation with the not-yet-updated `m_done`. The would lead to flushes like:

```
chunks: ['foo', 'bar', 'baz'], code: 0, gen !== null
chunks: [''], code: 200 gen === null
```

This diff just changes how we read the `m_code` field to match our understanding of the API better.

Reviewed By: jano

Differential Revision: D13951555

fbshipit-source-id: 7c7b47d418cf192a70f523c528b566d8e79af171