Add nast_visitor module
Summary:
In order to not depend on the `check_happly` and `tenv` environment fields, I've added a sort of starter / dummy env to the nast_visitor itself. Most of the information stored in the `NastCheck.env` we can pull from the classes, methods and tyepdefs directly.

This will make it so we can start moving other checks into `nast_checks` without being blocked by the tparam constraints.

Reviewed By: vassilmladenov

Differential Revision: D14070101

fbshipit-source-id: 9b14950947d26ea5102fb48a101ef0122c5887ac