simplify unions in hovered types
Summary:
With new inference, since type variables are not solved right away, we construct unions of type variables like (#1 | #2) which end up appearing as silly unions like (string | string) in the final tast. With this diff, we run the expander in Tast_expand on types that are being hovered, and we make this expander also simplify unions.

NB: I also considered doing that on the full tast at the end of type_file in Typing_check_utils, but that damaged the perf too much.

Reviewed By: Wilfred

Differential Revision: D14131488

fbshipit-source-id: 9ef1d2499d017b26678773d54b720477a99617a5