Fix autocomplete, outline, go-to-defintion for .hack files that contain `<?`
For example:
```
function do_some_codegen(): void {
  file_put_contents('foo.txt', '<?hh');
  //                            ^ this is not the start of code, this is
  //                            a string literal. If it weren't a .hack file,
  //                            it would be the start of code
}
```

Right now, how this is parsed entirely depends on the file extension.
Various parts of the IDE support create instances of the parser with
dummy filenames, so "is this a .hack file?" isn't a working approach.

This is not a good fix - but it is a small fix that I /think/ is
compatible with existing usage, and is suitable for committing to a `.z`
release.

This PR:
- makes several IDE entrypoints to the parser effectively set the `force_hh` flag
- changes the lexer to look at a global byref flag for 'ide mode' - if so, we assume it's a hack file,  and `<?` must either be the first line, or the second line, with the first line being a shebang line. If neither's two, we parse it as a .hack file.

This is based on two assumptions:

non-.Hack Hack files are either of this form:

```
<?hh
foo();
```

... or this form...

```
<?hh
```

Definitely not of the form:

```
Foo. please echo this content.
<?hh
```

... which is valid PHP.

The second assumption is that IDE utiltities
are invoked for real PHP files.