Remove emit_expr's last_pos argument
Summary:
This argument is used only when emitting locals and could be used to override
positions of opcodes that load locals. It was also used by other types of
expressions, but I simplified them in the past. The primary existence of this
argument seems to be exact position compatibility with emitter.cpp.

This matters only when a local is not set, causing notice or exception to be
thrown. In these cases, it's better to point to the actual local rather than
the encapsulating expression, so this should improve error messages while
making code simpler.

Reviewed By: oulgen

Differential Revision: D14188541

fbshipit-source-id: 67464589a9332b250955957ae3153208fa8e834d