Make DebuggerSerialize format serialize native data.
Summary:
We never called Native::nativeDataSleep or added its return value to the serialization output when the serialization format was DebuggerSerialize.
However, in the unserializer, in order to avoid having uninitialized native data, we always call Native::nativeDataWakeup, even if the format is DebuggerSerialize and even if the serializedNativeDataKey is not present in the serialized data.
This causes problems in hphpd:
```
localhost> p new ReflectionClass('stdClass')
wire error: unserialize() threw 'Native data of ReflectionClass should be a class name
'
null
```

Reviewed By: paulbiss

Differential Revision: D14192043

fbshipit-source-id: 877169e3491610ad3113e13725266452d59d190d