Have the ReifiedName instruction take the name as an immediate
Summary:
Now that ReifiedName is always taken as a literal string and also we do not need to know whether the name is for a class or for fun/meth, lets have the bytecode take the name as an immediate.

This simplifies code shared between this instruction and `RecordReifiedGeneric`. On the next diff, I'll implement the JIT side of things and this makes it simpler.

Reviewed By: jano

Differential Revision: D14143204

fbshipit-source-id: 2a763c768fd07b6fe1424690dfe8631940d0177e