Add a flag controlling which HH_FIXMEs are to be ignored by the parser
Summary:
We would like to be able to add HH_FIXMEs for the errors discovered by the new type inference algorithm but don't want these prematurely removed by the `--remove-dead-fixmes` feature.

This diff adds a new flag, `ignore_fixme_regex`, that can be set in `.hhconfig` to control which HH_FIXMEs are to be ignored and not collected by the full fidelity parser.  If we set it, say, to `.*new_inference.*`, then on the one hand, HH_FIXMEs containing substring `new_inference` will not hide any errors; and on the other hand, the `--remove-dead-fixmes` feature will keep such HH_FIXMEs.

Reviewed By: andrewjkennedy

Differential Revision: D14123969

fbshipit-source-id: 0b3d61fda389b5cb859f436e51ae6c7a5884134c