remove EnableContextInErrorHandler
Summary:
providing the context as part of the error handler gives the handler access to variables in the frame.
at the top level, this context is `$GLOBALS` but without the typical COW-less behavior when `$GLOBALS` is passed to a function (as it's a copy).
at a function level, this is access to the locals.

error handler having access to frame locals is odd behavior to have outside of a debugger.

removal was announced in https://hhvm.com/blog/2018/12/17/hhvm-3.30.html

this is deprecated in php7.2 as well.

Reviewed By: ricklavoie

Differential Revision: D14157588

fbshipit-source-id: a2d027e977a540341d0311829f4a8e237e9aabc6