Localize where constraints when performing NAST checks on a function/method
Summary:
Currently, generic type parameters in functions and methods are localized twice: in `NastCheck` and `Typing`.  However, where constraints are only localized (and added to the environment) once in `Typing`.  As a result, the bounds generated by where constraints are not available in the environment when we check if the constraints on the generic type parameters arising from the type hints of function/method parameters are implied by the bounds on the generic type parameters specified in the function/method/enclosing class declaration.  An example will make it clearer:
```
<?hh // strict
// Copyright 2004-present Facebook. All Rights Reserved.

class A {}

class Test<Tc as A> {
  public function f<Tf as Tc>(Test<Tf> $t): void {}

  public function g<Tg>(Test<Tg> $t): void where Tg as Tc {}
}
```
Here `f` type-checks while `g` produces an error:
```
File "/home/manzyuk/fbsource/fbcode/hphp/hack/test/where.php", line 9, characters 25-32:
Some type constraint(s) here are violated (Typing[4110])
  File "/home/manzyuk/fbsource/fbcode/hphp/hack/test/where.php", line 6, characters 12-13:
  'Tc' is a constrained type parameter
  File "/home/manzyuk/fbsource/fbcode/hphp/hack/test/where.php", line 6, characters 18-18:
  This is an object of type A
  File "/home/manzyuk/fbsource/fbcode/hphp/hack/test/where.php", line 9, characters 30-31:
  It is incompatible with a value of generic type Tg
```
At the time the constraint `Tg <: A` arising from the instantiation `Test<Tg>` is checked we don't have any additional information in the environment about `Tg` because the constraint `Tg <: Tc` hasn't been added and we fail to accept this instantiation as valid.

An easy fix is to localize where constraints in `NastCheck` too.  I'm not very familiar with `NastCheck` and am not sure why there is this duplication between `NastCheck` and `Typing`.  I'm open to better ways to solve this issue.

Reviewed By: MaggieMoss

Differential Revision: D13987844

fbshipit-source-id: 5539109a5154bbc7e7a162951ea44b3c66aa0def