Updating AST with Pocket Universes syntax
Summary:
Update of the AST to deal with Pocket Universes notations:
- new constraint `from`
- new expression `Atom` (which is lexed as `:name`) at the moment

Necessary update to naming/ and typing/ are also here in the following
states:
- naming: I'll update Aast in a subsequent commit so parts of ast_to_nast.ml are missing for now.
- typing: typing will be part of a future commit, once parsing is done.

The feature is guarded by an experimental flag (off by default) and can be
turned on using a command line flag `--pocket-universes` in `hh_single_parse`.
For parsing to succeed, the file header must be `<?hh // experimental`.

Reviewed By: oulgen

Differential Revision: D13859680

fbshipit-source-id: c7a894b4cb788c69baf6a27adfa55778cf400ff0