Add logic for type params marked with __Newable
Summary:
In order for us to be able to typecheck `new T($a, $b)`, we need to be able to guarantee the signature of the constructor for `T` over a range of classes. We have 3 options for parametrizing T then:
- **A final class**. This cannot have subclasses that change the constructor parameter types, so it is safe
- **A class that has a final constructor**, either from the class itself or inherited. Again, the constructor cannot be overridden by a subclass.
- **A class that has a consistent constructor**, either by having `<<__ConsistentConstruct>>` or a parent that has `<<__ConsistentConstruct>>`. The attribute ensures that subclasses will have the same constructor signature.

Here is an example bad case that shows off this requirement.
```
class A {}
class B extends A {
  final public function __construct(int $i): void {
    parent::__construct();
  }
}
class C extends A {
  final public function __construct(string $s): void {
    parent::__construct();
  }
}

function f<<<__Newable>> reify T as A /* bad */>(): void {
  new T(); // from A
}

function g(): void {
  f<B>(); // fails at runtime
  f<C>(); // same
}
```
Even though classes `B` and `C` are both valid and fit the constraint, they don't give the desired behavior. This is why the type parameter `T` itself must be constrained by an appropriate class that satisfies the above conditions.
```
function g<<<__Newable>> reify T as B>(): void {
  new T(42);
}
function h<<<__Newable>> reify T as C>(): void {
  new T("hello");
}
```
Interfaces are not considered by this diff, as they are currently broken and provide no guarantee of constructor consistency in subclasses.
```
interface I {
  public function __construct(string $s);
}
class C implements I {
  public function __construct(string $s) {}
}
class CC extends C {
  public function __construct(int $i) {
    parent::__construct("hello");
  }
}
```

Reviewed By: oulgen

Differential Revision: D13737106

fbshipit-source-id: c98fc8c85ffe904e80f99c6820635c0d447a12e1