Improve error messages produced by the new inference algorithm for invariant generic classes
Summary: A subtyping proposition `Inv<t1> <: Inv<t2>`, where `Inv` is an invariant generic class, and `t1` and `t2` are nonequivalent types, is unsatisfiable.  The error message produced in this case by the legacy algorithm contains a helpful hint, e.g.: "This is type `t1` ... It is incompatible with type `t2` ... **Considering this type argument is invariant with respect to `Inv`**".  In the new type inference algorithm the highlighted part is not always produced.  This diff attempts to fix this.

Reviewed By: andrewjkennedy

Differential Revision: D14241804

fbshipit-source-id: 27314403c7f6aaf706cb44b40d97f0cbd48fbc6f