Implement ClsRefGetTS in the JIT
Summary:
This diff implements most of ClsRefGetTS in the JIT and calls interpone for the part it doesn't.
```
class C {}
class D<reify T> {}

function f<reify T>() {
  new T();
}

f<C>();      // implemented in the JIT
f<D<int>>(); // interpone
```

First case it likely to happen much more often than the second so I only chose to implement that case but if there is hot code in the future, we can implement the rest too.

This diff also cleans up `ClsRefGetC`.

Reviewed By: ricklavoie

Differential Revision: D14232939

fbshipit-source-id: 81ead1d4181b6999a3a52514895aabe98118c5db