Updating Parser with support for pocket universes
Summary:
Now that current AST is updated, the parser needs to be updated to output
the correct AST tokens for PU structures.

The feature is guarded by an experimental flag (off by default) and can be
turned on using a command line flag `--pocket-universes` in `hh_single_parse`.
For parsing to succeed, the file header must be `<?hh // experimental`.

Reviewed By: Wilfred

Differential Revision: D14022507

fbshipit-source-id: 8ae2e234056fde3b555fdf2882bd227740f3c98e