Update type of Aast.AwaitAll
Summary:
The AAST is more permissive in its representation of AwaitAll than the AST, allowing a list of `expr option * expr`, while the AST only allows a list of `id option * expr`.

As AAST.AwaitAll is created from the AST's version, we can change the argument type of AAST.AwaitAll to `((lid option * expr) list)`.

This also makes it easier to convert the codegen modules to use AAST.AwaitAll.

Differential Revision: D14065896

fbshipit-source-id: f3224adf136883b64cb703244cd38cf4f02df84b