Updating Aast with support for pocket universes
Summary:
Now that the parser is updated, Aast needs to be updated with
the PU structure.

Necessary update to ai/ naming/ and typing/ are also here in the following
states:
- ai: some part I'm not sure how to update at the moment. Since the feature
  is boxed by flags, I'll get in touch with the relevant people once the
  syntax is finalized
- naming: update of aast and related files
- typing: typing will be part of a future commit, once ast/parsing is done.

The feature is guarded by an experimental flag (off by default) and can be
turned on using a command line flag `--pocket-universes` in `hh_single_parse`.
For parsing to succeed, the file header must be `<?hh // experimental`.

Reviewed By: Wilfred

Differential Revision: D14023048

fbshipit-source-id: 2ade0e452fb15c0c06430cfe2d85e328fef10e5a