New inference: flag eagerly-solved type variables and implement lint rule for type suggestion
Summary: Add a new entry to `tyvar_info` that records whether or not a type variable has been subject to an attempt at eager solving, in order to check a construct such as method invocation or array access. If this is the case, when expanding the type, produce a special reason code `Rsolve_fail` that can be used by linters (and, in future, code mods) to suggest a type hint.

Reviewed By: manzyuk

Differential Revision: D14131273

fbshipit-source-id: c1e788bebf377abf3aa585153a6b41293f2eb6b6