Type parameter 2 of preg_grep as `varray_or_darray`
Summary:
`preg_grep` will preserve the keys for the output, so we can't type the parameter as `varray`; at the same time, I don't think it's reasonable to require a `darray` to be passed in. So, `varray_or_darray` it is.

The cpp type needed to be changed to `Variant` to match `varray_or_darray` (https://fburl.com/wdmf03qn)

Reviewed By: billf

Differential Revision: D14143060

fbshipit-source-id: 9ee09f19fefeb2572c76b6d140e2f163cd54ce65