Add RuntimeOption::EnableReifiedGenerics to parser flags
Summary:
Abstract type constants with illegal generics can no longer be used in WWW and the codebase is cleaned up. We can remove this check now.

That check is the only place where we check in runtime that uses RuntimeOption::EnableReifiedGenerics, so we can add it to the parser flags now.

Reviewed By: vassilmladenov

Differential Revision: D14235659

fbshipit-source-id: d82ab8bb866b88b92b8d43d63f61d708af28f51a