Use a single chm for contextualReturnTypes
Summary:
Using D14291046 I noticed that the Index was allocating an enormous
amount of memory in a single std::vector (*just* for the vector, not
including any memory allocated separately by its elements).

Looking more closesly, it turned out that the tbb::chm in FuncInfo was
~670 bytes, and we allocate a FuncInfo for every php::Func.

We really don't need one chm per FuncInfo; we can simply include the
Func in the key.

Reviewed By: ricklavoie

Differential Revision: D14291251

fbshipit-source-id: c650a64f39a5aa9664dd60bf3b0f50fda3fd31c4