Improve error messages produced by the new inference algorithm when type-checking stringish objects
Summary: The old type inference algorithm produces an error "You cannot use this object as a string", whereas the new algorithm produces a generic error "This is an object of type Stringish ... It is incompatible with an object of type C".  I don't know how desirable it is to provide the former rather than latter error message, but here is an idea for a small refactoring that might prove useful in other situations where we would like to produce better error messages: let `sub_type` take an optional argument, a function that is called on a failure to simplify a subtyping constraint between two types, passing as arguments sub- and supertype arising on recursive calls to `simplify_subtype` that are patently incompatible.  (Currently, we are using `Typing_utils.uerror` for this; basically, I'm suggesting to pass an error reporting function as an optional parameter.)

Reviewed By: andrewjkennedy

Differential Revision: D14242965

fbshipit-source-id: b6883cb0ddf26517b43c11e357264644d07a7f7f