New inference: apply narrowing to type variable bounds for property access
Summary:
Currently, when we encounter `$x->f` and `$x` has a type variable type, we simply "solve" the type variable, typically to its lower bounds, in order to type check the property access. Unfortunately this can be too constraining, e.g. if the lower bound is `exact C`, or the field is inherited from a base class. Ideally we would introduce new type or constraints of the form `type t has property f of type u`, but this is a major change.

An alternative is to "narrow" the bounds on the type variable. For example, if we know
```
exact C <: #1
```
then it suffices to add a bound
```
exact C <: #1 <: B
```
if class `B` is the base class where property `f` was defined.

In future we will extend this technique to method invocation, and array access and other overloaded operations, hence the introduction of a generic helper `expand_type_and_narrow`.

Reviewed By: manzyuk

Differential Revision: D14258242

fbshipit-source-id: 69f92a2ace4991884373e22825d0bf9f408cb669