New inference: eager solve for null refinement only when lower bounds have explicit null
Summary:
We are getting a lot of false positives for null refinement (`$x !== null`) for variables whose type cannot be resolved. Instead of rejecting such code, it seems better just to avoid refining the type. Usually there will be an error down the line anyway (because we try and invoke a method, or whatever). So we now only "eager solve" type variables in such cases when one of the lower bounds is an explicit `null`, or `?t`, or has `?t` as a bound.

Longer term, we will implement null refinement by intersection with `nonnull`.

Reviewed By: kmeht

Differential Revision: D14278771

fbshipit-source-id: 97efe343465c219b70b2d638a37fc1e795cb929f