Remove \HH\asm from systemlib
Summary: We want to remove inline asm from HHVM since it is currently only used in 3 places and the uses cases are not strong enough to justify the heavy lifting it brings. This diff converts those use cases to intrinsic functions. Killing inline asm would clean up a lot of code that is only used to support this.

Reviewed By: alexeyt

Differential Revision: D14288560

fbshipit-source-id: 2bf8cb31d7b9484ecd2c2933cbcf9aafdc30ac22