Fix HHBBC crash when CheckPropTypeHints=3
Summary:
We can't look at class or type aliases when emitting units because they get
freed early to reduce memory pressure. Index::lookup_public_static() may try to
look at type aliases in order to resolve property type-hints, and we call that
during unit emission. Fix this by not trying to adjust the "unknown" portion of
the static's type according to the static's type-hint. This in theory makes the
types worse, but in practice does not seem to make any difference at all.

Reviewed By: jjgriego

Differential Revision: D14319472

fbshipit-source-id: c263459856baad62c379a9dbdb2729e3522c60b4