Remove assume_php setting
Summary:
The `assume_php` setting is a feature originally added to aid in gradual typing. When Hack encountered a class name which was not defined in any known Hack file, this setting would cause Hack to proceed as if that class existed in some PHP source file not visible to the typechecker. In a world where we are aiming for complete, sound type coverage, there isn't really a place for such a setting. Last August, we announced our intention to remove `assume_php` and turned it off by default.

Since D13551879, this setting is always disabled in all known real-world repositories and use cases. Historically, in addition to being an aid for gradual typing, this setting was used as an error suppression mechanism for IDE features. When writing D13551879, we found that this was no longer necessary, and the flipping of the `assume_php` flag in IDE logic was removed in that diff. This removed the last scenario where this flag was enabled, and it seems that it is now safe to remove it.

Many test cases are affected by the removal of this behavior, since it was enabled by default in some of our test suites:

- Some tests used unknown class names as a shorthand for creating a dummy class type or function in tests exercising other behavior (e.g., tests exercising array behavior might perform array operations on an `array<UnknownClass>`). In these cases, I defined the class type or function explicitly.
- This behavior masked typos in some tests. I have attempted to fix the typos in these cases.
- Some tests explicitly exercised the assume_php behavior, or its interaction with other features. I either removed these tests altogether, or added the missing name error to their error output (when their output involved printing multiple errors).
- The reactivity tests reassign_in_foreach3.php and reassign_in_foreach4.php did not correctly refer to HH\Rx\AsyncIterator. Correcting them caused another error--the Typing_reactivity module does not permit using foreach to iterate over an instance of HH\Rx\AsyncIterator. It seems to me that this is intended to be permitted, so I changed Typing_reactivity to allow it.

Reviewed By: Wilfred

Differential Revision: D14037113

fbshipit-source-id: 61eb28c2432b6d2a94f7e046b767e28d3bd51882