Avoid examining folded methods table when declaring condition_types
Summary:
In order to look up methods lazily using the linearization, it is necessary for us to do away with the folded collection of methods `Decl_inherit.ih_methods`. Decl currently depends on examining this collection to build up its set of condition types, a set of class names (provided by reactivity-related attributes) on which the declared class needs to add a dependency.

This diff changes condition_types to be built up during Decl by looking up the shallow declarations of each method in the class's ancestors (via `dc_ancestors`, AKA `impl`), so that it is not necessary to examine the folded collection of methods during class declaration.

This involves a small behavioral change--if a method with a conditionally reactive attribute was overridden by one without, previously, we would not add the type referred to by the attribute to condition_types, since the method with the attribute is no longer visible in the folded methods table after being overridden. I don't think this is harmful (it only causes more fanout when the class referred to in the conditionally-reactive attribute is changed).

In the future, it will be possible to use the linearization to look up condition_types from inherited methods, reproducing the old behavior. I'm not sure that this will be necessary, though (Decl_compare will need to work differently anyway, so we may end up not needing to replicate the old behavior). In the meantime, the migration to lazy method lookup is easier if condition_types does not rely on a folded methods table.

In order to use `Shallow_classes_heap.get` (which makes use of `Parser_heap`) in Decl, it was necessary to change the `Unit_hack` test to add the test file to the parser heap.

Reviewed By: vladima

Differential Revision: D13568133

fbshipit-source-id: 775307ad93615c6affbc45211989f662b00f83c7