Reorder MRO to match Decl_inherit
Summary: I think the original intent in Decl_linearize was to match Decl_inherit (hence the copy-paste of from_parent), but it looks like the ordering isn't quite right. This may have been intentional (maybe it better matches the behavior of the runtime?), but whether it was or not, for now I would like the MRO to model the typechecker's historical behavior as closely as possible.

Reviewed By: arxanas

Differential Revision: D13579468

fbshipit-source-id: b1c4b7873cc424b64a19e64372a8cb8aec1962ed