Don't crash on dependency cycles in Decl_linearize
Summary:
When we made Decl_linearize independent of Decl ({D13132772}), it was necessary to keep track of the stack of classes we were linearizing in order to avoid trying to include a class linearization within itself (causing either a stack overflow from infinite recursion or a Lazy.Undefined exception from Sequence.memoize).

Unfortunately, this is not sufficient. It is still entirely possible to get Decl_linearize to raise Lazy.Undefined. Consider these class definitions:

```
class A extends B {}
class B extends A {}
```

Suppose we look up the linearization for B. Decl_linearize will cache, for B, a lazy sequence which, when advanced, will emit `B`, then the linearization of B's ancestor A (which is not yet computed or cached). Suppose we then look up the linearization for A. Decl_linearize will again cache, for A, a lazy sequence which will emit `A`, then the linearization of A's ancestor B.

Suppose we then read the first element of B's linearization (using `Sequence.next`), and then attempt to iterate over the linearization of A.

The linearization of A will first yield the Child mro_element for A. It will then proceed to look up the linearization of A's ancestor B, which is in cache (and whose first element is memoized due to our use of `Sequence.next` above). It will begin iterating over B's linearization and yielding its elements. The first element, B, is memoized, and is yielded as part of A's linearization.

Next, we continue iterating over B's linearization to yield its elements as a part of A's linearization. The next element is from the linearization of B's ancestor, A. This linearization is in cache, and A is not in the class_stack used by B's linearization (since it was created in a separate toplevel linearization lookup), so we are now emitting A's linearization as part of A's linearization (despite our efforts to avoid this with class_stack). This doesn't cause a crash right away--we will happily proceed with emitting A as an element of A's own linearization. The crash comes once we have advanced the inner linearization of A to the point where it is computing the same element as the outer linearization of A, at which point `Lazy.Undefined` is raised.

[This diff](differential/diff/67824781/) adds some debug logging to demonstrate this example. Here is its output:

```
\A not in cache
\B not in cache
Emitting Child for \B
Advanced B's linearization; emitted \B
Linearization for class \A:
Emitting Child for \A
Got next element in linearization of \A: \A
Next_ancestor for \A is \B
\B in cache
Next element for \A
Next ancestor element for \A is \B (emitted)
Got next element in linearization of \A: \B
Next element for \A
Next_ancestor for \B is \A
\A in cache
Next element for \B
Next ancestor element for \B is \A (emitted)
Next ancestor element for \A is \A (emitted)
Got next element in linearization of \A: \A
Next element for \A
Next element for \B
Next ancestor element for \B is \B (emitted)
Next ancestor element for \A is \B (skipped)
Next element for \A
Next element for \B
Next ancestor element for \B is \A (skipped)
Next element for \B
Uncaught exception:

  Lazy.Undefined
```

Reviewed By: arxanas

Differential Revision: D13570821

fbshipit-source-id: f2859c1e6819369d1ba90b62cd8a56078393451c